Fix: 전시회상세페이지 북마크 실패해도 아이콘 변경되는 오류 수정
